Fix: fix LoaLoan and findCreditCardActivityBetweenDates
update readme
fix findCreditCardActivityBetweenDates in CrdCreditCardController and CrdCreditCardService
add GenCustomizedResponseEntityExceptionHandler
fix LoaLoanController
fix LoaCalculateLateFreeResponseDto
remove LoaPayInstallmentDto
add LoaPayLoanOffResponseDto
remove LoaPayOffDto
add new error messages to LoaErrorMesage
fix LoaLoanMapper
fix LoaLoanService
fix LoaLoanValidationService
fix CrdCreditCardController

+++ src/main/java/com/cbarkinozer/onlinebankingrestapi/app/loa/dto/LoaCalculateLateFeeResponseDto.java
@@ -1,7 +1,6 @@
 package com.cbarkinozer.onlinebankingrestapi.app.loa.dto;
 
 import lombok.Data;
-
 import java.math.BigDecimal;
 
 @Data
@@ -10,7 +9,7 @@ public class LoaCalculateLateFeeResponseDto {
     private BigDecimal totalLateFee;
     private BigDecimal lateFeeRate;
     private BigDecimal lateInterestTax;
-    private Integer lateDayCount;
+    private Long lateDayCount;
 
 
 }
